refactor: move related functions
